+++ src/controller/startLottoGame.js
@@ -1,6 +1,6 @@
-import LottoMachine from "../domain/LottoMachine.js";
+import LottoMachine from "../domain/LottoMachine/LottoMachine.js";
 import OutputView from "../view/OutputView.js";
-import profitCalculator from "../domain/profitCalculator.js";
+import profitCalculator from "../domain/profitCalculator/profitCalculator.js";
 import generateAnswerLotto from "../domain/generateAnswerLotto.js";
 import InputHandler from "./inputHandler.js";
