+++ src/domain/profitCalculator/calculateProfitRate.js
@@ -1,4 +1,4 @@
-import { HUNDRED_PERCENT } from "../constants/constant.js";
+import { HUNDRED_PERCENT } from "../../constants/constant.js";
 
 const calculateProfitRate = (profitAmount, purchaseAmount) => {
   return (profitAmount / purchaseAmount) * HUNDRED_PERCENT;
